Removed all unused code
